requirements.txt
torch
opencv-python
numpy
Pillow
gradio==4.28.3